* quaternion/dual_quaternion: implemented relational ops (==, !=, <=>)  * vector lib: implemented <=> op  * vr_context: moved flags in pose_t into an actual flags_t struct + made pose_t and flag_t comparable + updated users  * core: fixed cast in unix_timestamp_ms()/unix_timestamp_us() ...

+++ core/core.cpp
@@ -520,11 +520,11 @@ uint32_t unix_timestamp() {
 }
 
 uint64_t unix_timestamp_ms() {
-	return (uint32_t)chrono::duration_cast<chrono::milliseconds>(chrono::system_clock::now().time_since_epoch()).count();
+	return (uint64_t)chrono::duration_cast<chrono::milliseconds>(chrono::system_clock::now().time_since_epoch()).count();
 }
 
 uint64_t unix_timestamp_us() {
-	return (uint32_t)chrono::duration_cast<chrono::microseconds>(chrono::system_clock::now().time_since_epoch()).count();
+	return (uint64_t)chrono::duration_cast<chrono::microseconds>(chrono::system_clock::now().time_since_epoch()).count();
 }
 
 uint32_t get_hw_thread_count() {
